issue #312: Check volume group exists while creating a logical pool
When creating a logical storage pool, a volume group is created with the same
name.
And when deleting the storage pool the volume group keeps on the system. Which
is the correct behavior as the storage pool is only a concept and deleting it
should not affect the file system.
To avoid problems while creating a pool with an existent volume group name,
verify if a volume group exists with the given pool name and if so, warn the
user about it.

Also convert an error message parameter to string instead of virStoragePool
instance and update the po files as new messages were added.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>
Signed-off-by: Royce Lv <lvroyce@linux.vnet.ibm.com>