add timeout for sessions
The default timeout of sessions is 60 minutes.

Set the timeout of sessions 10 minutes explicitly.

Kimchi should have 10 minutes of time out value for the browser login
session.
If session got inactive for 10 minutes then it should expire
automatically.
And should ask user for relogin. This is required for the security
reason.

But this timeout will not take effect on some tabs, such as guest tab.
The root cause is because the front end refreshes the vm list every 5
seconds
by sending the "GET /vms" REST API call to the server.

The follow patch will solve this problem.

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>