Do not show a success msg when creating a template
When the user creates a new template, a green success message is
displayed on top of the screen. This behavior is not common in Kimchi;
when an operation performs successfully, nothing is shown. Actually, the
templates page is the only one that does that.

In order to make Kimchi more consistent, remove the code that displays a
success message when the user creates a new template.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>