Bug fix: Kimchi could not start guest with iscsi lun assigned as disk
There is an error in Kimchi when you try to start a guest created based
on a template that is using an iscsi storage pool. A lun should be used
as disk, but the iscsi case was not being expected and kimchi assigned
an .img file to the guest, raising error in the starting.
This patch fixed this problem.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>