Issue 294 The auto logout leaves action menu on page - add classes
When the user session ends, content that should no longer be visible
to the unauthenticated user is left in the DOM. The issue as written
notes that the guest action menu is displayed above the login window,
leading to a buggy user experience. Information about the users
guests, storage pools, networks, etc may also be left visible in the
browser.

This patch uses classes remove-when-logged-off and empty-when-logged-off
to mark those parts of the DOM that should no longer be available to
the user after the session has ended.

Signed-off-by: Adam King <rak@linux.vnet.ibm.com>