Issue 294 The auto logout leaves action menu on page - update JS
This patch updates those parts of the DOM marked with the classes
remove-when-logged-off and empty-when-logged-off
removing or emptying sections as noted by class.

Additionally, the change reloads the browser whenever the AJAX request
which caused us to detect that the session has ended is not marked for
resend. This handles the case where we are in some subwindow of the UI
when we detect that our session has ended, and loads the correct tab.

Signed-off-by: Adam King <rak@linux.vnet.ibm.com>