bug fix: Let package manager tool create the repository ID
The repository ID differs according to the repository type: deb or yum.
For yum repositories, the repository ID is kept in the repository
configuration file - so we can choose an arbitrary one as it will be
saved in a file.
But for a deb repository the ID does not exist. For internal proposals, we
build the ID for it based on url + dist + comps.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>