bug fix: Lock yum/apt operations
Just one yum/apt process can run per time so we need to lock those
operations to avoid race condition problems.

yum has a built in lock but it does not work well in kimchi as it only
prevents the user to run two different YUM applications. As the kimchi threads
are part of the same process, YUM does not really acquire the lock and
then we get the error:

[...]
  File "/usr/lib/python2.7/site-packages/urlgrabber/grabber.py", line 1258, in __init__
    self._do_open()
  File "/usr/lib/python2.7/site-packages/urlgrabber/grabber.py", line 1588, in _do_open
    self._set_opts()
  File "/usr/lib/python2.7/site-packages/urlgrabber/grabber.py", line 1375, in _set_opts
    self.curl_obj.setopt(pycurl.NOPROGRESS, False)
error: cannot invoke setopt() - perform() is currently running

So create a kimchi lock to solve this problem.
This lock can be used in any other place when needed.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>