bug fix: Display update progress on real time
When updating the system, the user should be able to see the progress to know
what is being done in the system.

stdout.read() will wait until get the completed output to return, ie, the user
will see the progress only when the update finishs.
So use stdout.readline() to read line by line and display the info to
the user.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>