Show error message from debug report generation async task
If the process of generating a debug report in Kimchi fails while it
was executing (e.g. due to running out of disk space to generate the
report), the error message was not being shown in the Kimchi UI.
Basically src/kimchi/control/tasks.py was putting the error message
in a JSON field called message, while the UI was looking for the error
message in a field called reason. 'reason' is used in standard
exception json returned only.

Signed-off-by: Leonardo Garcia <lagarcia@br.ibm.com>
Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>