Logical Storage Pool: Detect and Enumerate Multipath Block Devices
When creating a logical storage pool Kimci ignores multipath block
devices. This is on purpose because at the time the kimchi.disks was
implemented, there was not enough time to test every kind of devices.

This patch adds the missing support for multipath block devices. After
this patch, the kimchi.disks will include multipath devices in the
available partition list. The front-end code is also changed
accordingly.

v2
Use the major device number to determine if a deivce is a multipath one.
Use indexOf of the array object to determine if an element is in array.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>