Plugins UI: Correctly Load Plugin Tabs
In ui/js/src/kimchi.main.js we fetch all plugin tabs and concat them to
the tabs array. However array.concat does not work in place, it creates
a new array to store the result, so it does not correctly update the
tabs array. The plugin tabs are not loaded at all.

This patch fixes the problem by using tabs.push.apply(tabs, pluginTabs) .

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>