Fix hardcoded storage bus assignment in vmstorage
There is an error when a user tries to change the midia path in "Manage
Midia" at Guest tab. If the midia was created with SCSI bus, Kimchi is
going to try replace it by IDE bus, causing the following error:

No device with bus 'ide' and target 'sdc'

Bus IDE was hardcoded. This patch retrieves the original midia bus and
pass the value as parameter in order to create the right XML.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>