Fix get_storageservers API and UI retrieval of storage servers.
This issue was discovered when it was realized that the list of NFS
servers currently on the system was not being displayed in a drop-down
for the server IP/name while adding a new NFS server.

The UI was calling the getstorageservers API and passing a filter of netfs, but
the StorageServers class was referencing a non-existent object. This threw an
exception that was not passed up as an error, and an empty list was returned
every time.

Additionally, a change to the STORAGE_SOURCES dictionary was causing errors
when the StorageServersModel::lookup() tried to access a non-existent key in
in the new scsi value. This is seen if a scsi pool exists on a system.

To allow the API to work, and sucessfully retrieve only *server* objects, this
patch makes the following changes:

1. Add a STORAGE_SERVERS list to storageservers.py containing only valid
   server types.
   The API doc states that currently only netfs is supported.
2. Since storagetargets correlates to an individual server (see the
   API doc for storagetargets) use the new STORAGE_SERVERS list instead
   of the sources ones.
3. Move the STORAGE_SOURCES dictionary into storagepools.py, since it should
   cover all pool types. In the future, it should be expanded to include
   iscsi, netfs, dir, scsi fc_host, etc.
4. Updates mockmodel to incorporate the new STORAGE_SERVERS list.

To test, create different types of pools, one nfs, and use:

$ curl -k -u user:pass -H 'Content-type: applicatiojson' -H 'Accept: application/json' https://localhost:8001/storageservers?_target_type=netfs

and

$ curl -k -u user:pass -H 'Content-type: applicatiojson' -H 'Accept: application/json' https://localhost:8001/storageservers

Or create at least one NFS pool, then when creating a second NFS pool, the
  names/IPs of the others will appear in the UI's server field called
  "NFS Server IP"

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>