Add/remove users and groups to VMs
To update (add/remove) the users and groups of a virtual machine, you
should use the REST command:

/vms/<vm_name> PUT "{ 'users': [ 'user1', 'user2' ],
                      'groups': [ 'group1', 'group2' ] }"

Users and groups associated with a virtual machine must be valid when
they are
added to the VM. Such verification is not done when they are removed
from the
system. This means that virtual machines may contain outdated
information about
users and groups.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>
Signed-off-by: Aline Manera <alinefm@br.ibm.com>