repository: Place checkbox to the left of its label
In general, it is more common to have a checkbox placed to the left side
of its corresponding label. Also, all the other checkboxes in Kimchi are
placed to the left side of their labels.

Change the location of the checkbox "Repository is a mirror" so that the
control element is placed to the left side of its label.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>