Define the default tab more clearly
In order to define the default tab, the developer needs to update the
code in more than one place. Also, the corresponding variable name does
not reflect what it actually means.

Define the default tab in a single variable and make all the other
values dependent on it.
Also, rename the variable from "firstTab" to "defaultTab".

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>