Github #368: Removing 'resend' tag from getHostStats JSON
As described in #368, the Hosts tab wasn't appearing after
relogging (due to restarting kimchid or even when the session
expires). The issue was in the getHostStats JSON, which was
marked with the 'resend' flag, preventing the Host tab from
reloading.

This behavior is set in the kimchi.login_window.js. This flag
isn't used for any other purpose, thus its removal does not
break the refresh of the System Statistics frame in the Hosts
tab. A description of the right usage of the 'resend' tag can
be found in the header of kimchi.api.js.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>